+++ packages/atomic/src/components/facets/atomic-timeframe-facet/atomic-timeframe-facet.tsx
@@ -28,7 +28,12 @@ import {FacetPlaceholder} from '../atomic-facet-placeholder/atomic-facet-placeho
 import {FacetContainer} from '../facet-container/facet-container';
 import {FacetHeader} from '../facet-header/facet-header';
 import {FacetValueLink} from '../facet-value-link/facet-value-link';
-import {BaseFacet, parseDependsOn, validateDependsOn} from '../facet-common';
+import {
+  BaseFacet,
+  parseDependsOn,
+  shouldDisplayInputForFacetRange,
+  validateDependsOn,
+} from '../facet-common';
 import {Timeframe} from '../atomic-timeframe/timeframe';
 import {FacetValueLabelHighlight} from '../facet-value-label-highlight/facet-value-label-highlight';
 import dayjs from 'dayjs';
@@ -40,6 +45,8 @@ import {
   FocusTargetController,
 } from '../../../utils/accessibility-utils';
 import {MapProp} from '../../../utils/props-utils';
+import {FacetValuesGroup} from '../facet-values-group/facet-values-group';
+import {randomID} from '../../../utils/utils';
 
 /**
  * A facet is a list of values for a certain field occurring in the results.
@@ -71,19 +78,24 @@ import {MapProp} from '../../../utils/props-utils';
   shadow: true,
 })
 export class AtomicTimeframeFacet
-  implements InitializableComponent, BaseFacet<DateFacet, DateFacetState>
+  implements InitializableComponent, BaseFacet<DateFacet>
 {
   @InitializeBindings() public bindings!: Bindings;
-  public facet?: DateFacet;
+  public facetForDateRange?: DateFacet;
+  public facetForDatePicker?: DateFacet;
+
   private dependenciesManager?: FacetConditionsManager;
   public filter?: DateFilter;
   public searchStatus!: SearchStatus;
   private manualTimeframes: Timeframe[] = [];
   @Element() private host!: HTMLElement;
 
-  @BindStateToController('facet')
+  @BindStateToController('facetForDateRange')
   @State()
-  public facetState?: DateFacetState;
+  public facetState!: DateFacetState;
+  @BindStateToController('facetForDatePicker')
+  @State()
+  public facetForDatePickerState?: DateFacetState;
   @BindStateToController('filter')
   @State()
   public filterState?: DateFilterState;
@@ -113,6 +125,10 @@ export class AtomicTimeframeFacet
    * Specifies if the facet is collapsed.
    */
   @Prop({reflect: true, mutable: true}) public isCollapsed = false;
+  /**
+   * The [heading level](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements) to use for the heading over the facet, from 1 to 6.
+   */
+  @Prop({reflect: true}) public headingLevel = 0;
   /**
    * Whether to exclude the parents of folded results when estimating the result count for each facet value.
    */
@@ -157,7 +173,7 @@ export class AtomicTimeframeFacet
     this.validateProps();
     this.manualTimeframes = this.getManualTimeframes();
     this.searchStatus = buildSearchStatus(this.bindings.engine);
-    this.manualTimeframes.length && this.initializeFacet();
+    this.initializeFacets();
     this.withDatePicker && this.initializeFilter();
     this.inititalizeDependenciesManager();
     this.registerFacetToStore();
@@ -170,7 +186,34 @@ export class AtomicTimeframeFacet
     this.dependenciesManager?.stopWatching();
   }
 
-  private initializeFacet() {
+  private initializeFacets() {
+    // Initialize two facets: One that is actually used to display values for end users, which only exists
+    // if we need to display something to the end user (ie: timeframes > 0)
+
+    // A second facet is initialized only to verify the results count. It is never used to display results to end user.
+    // It serves as a way to determine if the input should be rendered or not, independent of the ranges configured in the component
+    if (this.manualTimeframes.length > 0) {
+      this.initializeFacetForDateRange();
+    }
+    if (this.withDatePicker) {
+      this.initializeFacetForDatePicker();
+    }
+  }
+
+  private initializeFacetForDatePicker() {
+    this.facetForDatePicker = buildDateFacet(this.bindings.engine, {
+      options: {
+        numberOfValues: 1,
+        generateAutomaticRanges: true,
+        facetId: randomID(this.facetId || this.field),
+        field: this.field,
+        filterFacetCount: this.filterFacetCount,
+        injectionDepth: this.injectionDepth,
+      },
+    });
+  }
+
+  private initializeFacetForDateRange() {
     const options: DateFacetOptions = {
       facetId: this.facetId,
       field: this.field,
@@ -180,8 +223,8 @@ export class AtomicTimeframeFacet
       filterFacetCount: this.filterFacetCount,
       injectionDepth: this.injectionDepth,
     };
-    this.facet = buildDateFacet(this.bindings.engine, {options});
-    this.facetId = this.facet.state.facetId;
+    this.facetForDateRange = buildDateFacet(this.bindings.engine, {options});
+    this.facetId = this.facetForDateRange.state.facetId;
   }
 
   private initializeFilter() {
@@ -198,6 +241,9 @@ export class AtomicTimeframeFacet
   }
 
   private registerFacetToStore() {
+    if (!this.facetForDateRange) {
+      return;
+    }
     registerFacetToStore(this.bindings.store, 'dateFacets', {
       label: this.label,
       facetId: this.facetId!,
@@ -212,10 +258,14 @@ export class AtomicTimeframeFacet
   }
 
   private inititalizeDependenciesManager() {
+    if (!this.facetForDateRange && !this.filter) {
+      return;
+    }
     this.dependenciesManager = buildFacetConditionsManager(
       this.bindings.engine,
       {
-        facetId: this.facet?.state.facetId ?? this.filter!.state.facetId,
+        facetId:
+          this.facetForDateRange?.state.facetId ?? this.filter!.state.facetId,
         conditions: parseDependsOn(this.dependsOn),
       }
     );
@@ -278,10 +328,11 @@ export class AtomicTimeframeFacet
             this.filter?.clear();
             return;
           }
-          this.facet?.deselectAll();
+          this.facetForDateRange?.deselectAll();
         }}
         numberOfSelectedValues={this.numberOfSelectedValues}
         isCollapsed={this.isCollapsed}
+        headingLevel={this.headingLevel}
         onToggleCollapse={() => (this.isCollapsed = !this.isCollapsed)}
         headerRef={this.headerFocus.setTarget}
       ></FacetHeader>
@@ -343,7 +394,7 @@ export class AtomicTimeframeFacet
         isSelected={isSelected}
         numberOfResults={facetValue.numberOfResults}
         i18n={this.bindings.i18n}
-        onClick={() => this.facet!.toggleSingleSelect(facetValue)}
+        onClick={() => this.facetForDateRange!.toggleSingleSelect(facetValue)}
       >
         <FacetValueLabelHighlight
           displayValue={displayValue}
@@ -355,9 +406,11 @@ export class AtomicTimeframeFacet
 
   private renderValuesContainer(children: VNode[]) {
     return (
-      <ul part="values" class="mt-3">
-        {children}
-      </ul>
+      <FacetValuesGroup i18n={this.bindings.i18n} label={this.label}>
+        <ul class="mt-3" part="values">
+          {children}
+        </ul>
+      </FacetValuesGroup>
     );
   }
 
@@ -384,16 +437,20 @@ export class AtomicTimeframeFacet
   }
 
   private get shouldRenderValues() {
-    const hasInputRange = !!this.filterState?.range;
-    return !hasInputRange && !!this.valuesToRender.length;
+    return !this.hasInputRange && !!this.valuesToRender.length;
+  }
+
+  private get hasInputRange() {
+    return !!this.filterState?.range;
   }
 
   private get shouldRenderInput() {
-    if (!this.withDatePicker) {
-      return false;
-    }
-
-    return this.searchStatusState.hasResults || !!this.filterState?.range;
+    return shouldDisplayInputForFacetRange({
+      hasInput: this.withDatePicker,
+      hasInputRange: this.hasInputRange,
+      searchStatusState: this.searchStatusState,
+      facetValues: this.facetForDatePickerState?.values || [],
+    });
   }
 
   public render() {